Split up checker classes
