+++ src/mlx/warnings.py
@@ -10,8 +10,9 @@ import sys
 
 from pkg_resources import require
 
-from mlx.warnings_checker import (CoverityChecker, DoxyChecker, JUnitChecker, RobotChecker, SphinxChecker,
-                                  RobotSuiteChecker, XMLRunnerChecker)
+from mlx.junit_checker import JUnitChecker
+from mlx.regex_checker import CoverityChecker, DoxyChecker, SphinxChecker, XMLRunnerChecker
+from mlx.robot_checker import RobotChecker, RobotSuiteChecker
 
 __version__ = require('mlx.warnings')[0].version
